form_processing/requirements.txt
# Core requirements
streamlit==1.32.0
streamlit-drawable-canvas==0.9.3
pdf2image==1.16.3
pytesseract==0.3.10
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.2.0

# Text processing
python-Levenshtein==0.25.0
regex==2023.12.25

# OCR and Document Processing
easyocr==1.7.1
layoutparser==0.3.4

# Data handling
pandas==2.2.1
pydantic==2.6.3

# Development tools
pytest==8.0.2
black==24.2.0
flake8==7.0.0
mypy==1.8.0

# Optional - for enhanced document processing
doctr==0.7.1
torch==2.2.1
torchvision==0.17.1